requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
gradio>=4.0.0
rich>=13.0.0
streamlit>=1.30.0
